% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_GlobalFiler_3500_data.R
\name{get_GlobalFiler_3500_data}
\alias{get_GlobalFiler_3500_data}
\title{Convenience function that loads size regression and stutter files}
\usage{
get_GlobalFiler_3500_data()
}
\value{
Object of class list
}
\description{
Convenience function that loads size regression and stutter files
}
\examples{
gf <- get_GlobalFiler_3500_data()
}
