% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_kit_data.R
\name{get_kit_data}
\alias{get_kit_data}
\title{Get kit data such as loci, dye, sizes for most standard kits}
\usage{
get_kit_data()
}
\value{
Object of class list
}
\description{
Get kit data such as loci, dye, sizes for most standard kits
}
\examples{
# obtain the markers and dye colours for GlobalFiler
kit_data <- get_kit_data()
gf <- kit_data$GlobalFiler_Panel_v1
gf_markers <- unique(gf$Marker)
gf_colour_by_marker <- stats::setNames(gf$Color[match(gf_markers, gf$Marker)], gf_markers)
}
