% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_wide_table.R
\name{read_wide_table}
\alias{read_wide_table}
\title{Read wide table (.txt) with Allele1, Allele2, ... columns as is}
\usage{
read_wide_table(filename)
}
\arguments{
\item{filename}{Path to txt file.}
}
\value{
Dataframe
}
\description{
Read wide table (.txt) with Allele1, Allele2, ... columns as is
}
