% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_mixtures_with_relatives.R
\name{sample_mixtures_with_relatives}
\alias{sample_mixtures_with_relatives}
\title{Sample mixtures with random genotypes and random parameters according to priors}
\usage{
sample_mixtures_with_relatives(
  n,
  contributors,
  freqs,
  sampling_parameters,
  model_settings,
  sample_model,
  pedigree,
  results_directory,
  tag = "simulation"
)
}
\arguments{
\item{n}{Integer. Number of samples.}

\item{contributors}{Character vector with unique names of contributors. Valid names are "U1", "U2", ... for unrelated contributors or the names of pedigree members for related contributors.}

\item{freqs}{Allele frequencies (see \link{read_allele_freqs})}

\item{sampling_parameters}{List. Passed to the sample_model function.}

\item{model_settings}{List. Passed to the sample_model function.}

\item{sample_model}{Function such as \link{sample_log_normal_model}.}

\item{pedigree}{(optionally) \link[pedtools]{ped} object.}

\item{results_directory}{(optionally) Character with path to directory where results are written to disk.}

\item{tag}{Character. Used for sub directory name when results_directory is provided.}
}
\description{
Sample mixtures with random genotypes and random parameters according to priors
}
