% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allele_specific_stutter_model.R
\name{allele_specific_stutter_model}
\alias{allele_specific_stutter_model}
\title{Stutter model where the expected stutter rate depends on the allele and locus}
\usage{
allele_specific_stutter_model(stutter_types, size_regression)
}
\arguments{
\item{stutter_types}{List. See \link{stutter_type}.}

\item{size_regression.}{Function, see \link{read_size_regression}.}
}
\description{
Stutter model where the expected stutter rate depends on the allele and locus
}
\details{
When a pg_model is constructed (see \link{gamma_model}), a stutter model can optionally be applied.
}
