% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_internal_data_validation.R
\name{.validate_freqs}
\alias{.validate_freqs}
\title{Validate Allele Frequencies}
\usage{
.validate_freqs(freqs, required_loci = NULL)
}
\arguments{
\item{freqs}{List of allele frequencies.}

\item{required_loci}{Character vector of required loci names.}
}
\description{
This is an internal function to check the format and contents of the frequency list.
}
\keyword{internal}
