% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma_model.R
\name{gamma_model}
\alias{gamma_model}
\title{Defines a gamma model for peak height variability}
\usage{
gamma_model(
  mixture_proportions,
  mu,
  cv,
  degradation_beta = rep(1, length(mixture_proportions)),
  model_settings
)
}
\arguments{
\item{mixture_proportions}{Numeric vector}

\item{mu}{Numeric. Expectation of a full heterozygote contributing allele peak height.}

\item{cv}{Numeric. Coefficient of variation of a full heterozygote contributing allele peak height}

\item{degradation_beta}{Numeric Vector of same length as mixture_proportions. Degradation slope parameters for each contributor.}

\item{model_settings}{List. Possible parameters: \itemize{
\item locus_names. Character vector.
\item detection_threshold. Numeric vector (named) with Detection Thresholds. Defaults to 50 for each locus.
\item size_regression. Function, see \link{read_size_regression}.
\item stutter_model. Optionally a stutter_model object that gives expected stutter heights. See \link{global_stutter_model}.
}}
}
\description{
Defines a gamma model for peak height variability
}
\details{
Defines a gamma model as described by Bleka et al.
}
