% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kits.R
\docType{data}
\name{kits}
\alias{kits}
\title{Properties such as loci, dye, sizes for most standard forensic kits}
\format{
A list of data frames containing variables such as:
\describe{
\item{Panel}{Kit name}
\item{Marker}{}
\item{Allele}{}
\item{Size}{Fragment length}
\item{Color}{Dye colour}
}
}
\source{
\url{https://github.com/oyvble/euroformix/blob/master/inst/extdata/kit.txt}
}
\usage{
kits
}
\description{
A dataset containing the properties of forensic DNA kits.
}
\keyword{datasets}
