% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_normal_model.R
\name{log_normal_model}
\alias{log_normal_model}
\title{Defines a log normal model for peak height variability}
\usage{
log_normal_model(
  template,
  degradation = rep(0, length(template)),
  c2,
  stutter_model,
  size_regression
)
}
\arguments{
\item{template}{Numeric vector}

\item{degradation}{Numeric Vector of same length as template. Degradation parameters for each contributor.}

\item{c2}{Numeric. Allele variance parameter.}

\item{size_regression.}{Function, see \link{read_size_regression}.}

\item{stutter_model.}{Optionally a stutter_model object. See \link{global_stutter_model}.}
}
\description{
Defines a log normal model for peak height variability
}
\details{
Defines a log normal model as described by Bright et al.
}
