% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_contributor_genotypes.R
\name{sample_contributor_genotypes}
\alias{sample_contributor_genotypes}
\title{Sample genotypes for mixture contributors according to allele frequencies}
\usage{
sample_contributor_genotypes(
  contributors,
  freqs,
  pedigree,
  loci = names(freqs),
  return_non_contributors = FALSE
)
}
\arguments{
\item{contributors}{Character vector with unique names of contributors. Valid names are "U1", "U2", ... for unrelated contributors or the names of pedigree members for related contributors.}

\item{freqs}{Allele frequencies (see \link{read_allele_freqs})}

\item{pedigree}{(optionally) \link[pedtools]{ped} object}

\item{loci}{Character vector of locus names (defaults to names attr. of \code{freqs})}

\item{return_non_contributors}{Logical. Should genotypes of non-contributing pedigree members also be returned?}
}
\description{
Sample genotypes for mixture contributors according to allele frequencies
}
\details{
For each founder, a genotype is sampled randomly by drawing two alleles from allele frequencies.
}
\examples{

# read allele frequencies
freqs <- read_allele_freqs(system.file("extdata","FBI_extended_Cauc.csv",
                           package = "simDNAmixtures"))

# define a pedigree of siblings S1 and S2 (and their parents)
ped_sibs <- pedtools::nuclearPed(nch = 2,
father = "F", mother = "M",
children = c("S1", "S2"))

# sample genotypes for a mixture of S1 + U1 + S2
# where U1 is an unrelated person

sample_contributor_genotypes(contributors = c("S1","U1","S2"), freqs, ped_sibs)
}
