% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_gamma_model.R
\name{sample_gamma_model}
\alias{sample_gamma_model}
\title{Sample gamma model(s) with parameters according to priors}
\usage{
sample_gamma_model(number_of_contributors, sampling_parameters, model_settings)
}
\arguments{
\item{number_of_contributors}{Integer}

\item{sampling_parameters}{List. Needs to contain:
\itemize{
\item min_mu. Numeric of length one.
\item max_mu. Numeric of length one.
\item min_cv. Numeric of length one.
\item max_cv. Numeric of length one.
\item degradation_shape1. Numeric of length one.
\item degradation_shape2. Numeric of length one.
}}

\item{model_settings}{List. See \link{gamma_model}.}
}
\description{
Sample gamma model(s) with parameters according to priors
}
\examples{
gf <- get_GlobalFiler_3500_data()

sampling_parameters <- list(min_mu = 50., max_mu = 5e3,
                           min_cv = 0.05, max_cv = 0.35,
                           degradation_shape1 = 10, degradation_shape2 = 1)

model_no_stutter <- sample_gamma_model(number_of_contributors = 2,
                                      sampling_parameters = sampling_parameters,
                                      model_settings = gf$gamma_settings_no_stutter)

model_no_stutter$parameters

}
