% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_log_normal_model.R
\name{sample_log_normal_model}
\alias{sample_log_normal_model}
\title{Sample log normal model(s) with parameters according to priors}
\usage{
sample_log_normal_model(
  number_of_contributors,
  sampling_parameters,
  model_settings
)
}
\arguments{
\item{number_of_contributors}{Integer}

\item{sampling_parameters}{List. Needs to contain:
\itemize{
\item min_template. Numeric of length one.
\item max_template. Numeric of length one.
\item degradation_shape. Numeric of length one.
\item degradation_scale. Numeric of length one.
}}

\item{model_settings}{List. Possible parameters: \itemize{
\item locus_names. Character vector.
\item degradation_parameter_cap. Numeric.
\item c2_prior. Numeric of length two with shape and scale.
\item LSAE_variance_prior. Numeric of length one.
\item detection_threshold. Numeric vector (named) with Detection Thresholds. Defaults to 50 for each locus.
\item size_regression. Function, see \link{read_size_regression}.
\item stutter_model. Optionally a stutter_model object that gives expected stutter heights. See \link{global_stutter_model}.
\item stutter_variability. Optionally peak height variability parameters for stutters. Required when stutter_model is supplied.
}}
}
\description{
Sample log normal model(s) with parameters according to priors
}
\examples{
gf <- get_GlobalFiler_3500_data()

sampling_parameters <- list(min_template = 50., max_template = 1000.,
                            degradation_shape = 2.5, degradation_scale = 1e-3)

model_no_stutter <- sample_log_normal_model(number_of_contributors = 1,
                                            sampling_parameters = sampling_parameters,
                                            model_settings = gf$log_normal_settings)
}
