% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_log_normal_stutter_variance.R
\name{sample_log_normal_stutter_variance}
\alias{sample_log_normal_stutter_variance}
\title{Sample log normal stutter variance parameters according to priors}
\usage{
sample_log_normal_stutter_variance(log_normal_stutter_variability)
}
\arguments{
\item{log_normal_stutter_variability}{List of variability parameters. See \link{gf} for an example.}
}
\value{
Named numeric with stutter variance parameter for all stutter types. Names are \code{k2} concatenated with the name of the stuter type. See example.
}
\description{
Sample log normal stutter variance parameters according to priors
}
\examples{
gf <- gf_configuration()
log_normal_stutter_variability <- gf$log_normal_settings$stutter_variability
k2 <- sample_log_normal_stutter_variance(log_normal_stutter_variability)
}
