% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_log_normal_stutter_variance.R
\name{sample_log_normal_stutter_variance}
\alias{sample_log_normal_stutter_variance}
\title{Sample log normal stutter variance parameters according to priors}
\usage{
sample_log_normal_stutter_variance(log_normal_stutter_variability)
}
\arguments{
\item{log_normal_stutter_variability}{List of variability parameters. See \link{get_GlobalFiler_3500_data} for an example.}
}
\description{
Sample log normal stutter variance parameters according to priors
}
\examples{
 gf <- get_GlobalFiler_3500_data()
 log_normal_stutter_variability <- gf$log_normal_stutter_variability
 k2 <- sample_log_normal_stutter_variance(log_normal_stutter_variability)
}
