% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_many_pedigree_genotypes.R
\name{sample_many_pedigree_genotypes}
\alias{sample_many_pedigree_genotypes}
\title{Sample (many) genotypes for pedigree members according to allele frequencies by allele dropping and possibly taking linkage into account by simulating recombination.}
\usage{
sample_many_pedigree_genotypes(
  pedigree,
  freqs,
  loci = names(freqs),
  unrelated_names = character(),
  linkage_map,
  number_of_replicates = 1L,
  sex_locus_name = "AMEL",
  return_transmission_vectors = FALSE
)
}
\arguments{
\item{pedigree}{\link[pedtools:ped]{ped} object}

\item{freqs}{Allele frequencies (see \link{read_allele_freqs})}

\item{loci}{Character vector of locus names (defaults to \code{names} attribute of \code{freqs})}

\item{unrelated_names}{Character vector with names of any additional unrelated persons. Defaults to length zero.}

\item{linkage_map}{A linkage map specifying the recombination fractions between loci. If NULL, loci are assumed to be independent.}

\item{number_of_replicates}{An integer specifying the number of replicate genotype samples to generate. Defaults to 1.}

\item{sex_locus_name}{Character vector, defaults to "AMEL"}
}
\description{
Sample (many) genotypes for pedigree members according to allele frequencies by allele dropping and possibly taking linkage into account by simulating recombination.
}
\seealso{
\link{read_allele_freqs}
}
