% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_mixture_from_genotypes.R
\name{sample_mixture_from_genotypes}
\alias{sample_mixture_from_genotypes}
\title{Sample mixture profile with provided genotypes}
\usage{
sample_mixture_from_genotypes(genotypes, model, sample_name = "mixture")
}
\arguments{
\item{genotypes}{List of contributor genotypes. See \link{sample_contributor_genotypes}.}

\item{model}{pg_model object.}

\item{sample_name}{Character. Defaults to "mixture".}
}
\description{
Sample mixture profile with provided genotypes
}
\details{
A mixture profile is sampled according to the provided pg_model.
}
\examples{
# read allele frequencies
freqs <- read_allele_freqs(system.file("extdata","FBI_extended_Cauc.csv",package = "SimMixDNA"))

# define a pedigree of siblings S1 and S2 (and their parents)
ped_sibs <- pedtools::nuclearPed(nch = 2,
father = "F", mother = "M",
children = c("S1", "S2"))

# sample genotypes for a mixture of S1 + U1 + S2
# where U1 is an unrelated perso
genotypes <- sample_contributor_genotypes(contributors = c("S1","U1","S2"), freqs,
ped_sibs, loci = get_GlobalFiler_3500_data()$autosomal_markers)

}
