% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_pedigree_genotypes.R
\name{sample_pedigree_genotypes}
\alias{sample_pedigree_genotypes}
\title{Sample genotypes for pedigree according to allele frequencies}
\usage{
sample_pedigree_genotypes(pedigree, freqs, loci = names(freqs))
}
\arguments{
\item{pedigree}{Ped object}

\item{freqs}{Allele frequencies (see \link{read_allele_freqs})}

\item{loci}{Character vector of locus names (defaults to names attr. of \code{freqs})}
}
\description{
Sample genotypes for pedigree according to allele frequencies
}
\details{
For each founder, a genotype is sampled randomly by drawing two alleles from allele frequencies.
}
\examples{
freqs <- read_allele_freqs(system.file("extdata","FBI_extended_Cauc.csv",package = "SimMixDNA"))

ped_sibs <- pedtools::nuclearPed(nch = 2,
father = "F", mother = "M",
children = c("S1", "S2"))
sibs_genotypes <- SimMixDNA::sample_pedigree_genotypes(ped = ped_sibs,
freqs = freqs, loci = get_GlobalFiler_3500_data()$autosomal_markers)
}
