% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMASH_to_wide_table.R
\name{SMASH_to_wide_table}
\alias{SMASH_to_wide_table}
\title{Converts SMASH (SampleName, Marker, Allele, Size, Height) data to a wide table}
\usage{
SMASH_to_wide_table(x)
}
\arguments{
\item{x}{DataFrame.}
}
\value{
DataFrame
}
\description{
Converts SMASH (SampleName, Marker, Allele, Size, Height) data to a wide table
}
