% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stutter_type.R
\name{stutter_type}
\alias{stutter_type}
\title{Defines a stutter type to be used in an allele specific stutter model.}
\usage{
stutter_type(name, delta, stutter_regression, stutter_exceptions)
}
\arguments{
\item{name}{Character. Name of the stutter, e.g. "BackStutter"}

\item{delta}{Numeric. Repeat units gained (lost when negative).}

\item{stutter_regression}{Function. See \link{read_stutter_regression}.}

\item{stutter_exceptions}{Optinally a list. See \link{read_stutter_exceptions}.}
}
\description{
Defines a stutter type to be used in an allele specific stutter model.
}
\details{
When a pg_model is constructed (see \link{gamma_model}), a stutter model can optionally be applied.
}
\examples{
filename_bs_exceptions <- system.file("extdata","GlobalFiler_Stutter_Exceptions_3500.csv",package = "SimMixDNA")
bs_exceptions <- read_stutter_exceptions(filename_bs_exceptions)

filename_bs_regression <- system.file("extdata","GlobalFiler_Stutter_3500.txt",package = "SimMixDNA")
bs_regression <- read_stutter_exceptions(filename_bs_regression)
backstutter <- stutter_type(name = "BackStutter", delta = -1,
                            stutter_regression = bs_regression,
                            stutter_exceptions = bs_exceptions)
}
